+++ Sources/SQLite/SQLite.swift
@@ -90,41 +90,7 @@ public class SQLite {
             let count = sqlite3_column_count(statement.pointer)
 
             for i in 0..<count {
-                let name = sqlite3_column_name(statement.pointer, i)
-                
-                let column: String
-                if let name = name {
-                    column = String(cString: name)
-                } else {
-                    column = ""
-                }
-                
-                switch sqlite3_column_type(statement.pointer, i) {
-                case SQLITE_TEXT:
-                    let text = sqlite3_column_text(statement.pointer, i)
-                    
-                    var value: String = ""
-                    if let text = text {
-                        value = String(cString: text)
-                    
-                    }
-                    
-                    row.data[column] = .text(value)
-                    
-                case SQLITE_INTEGER:
-                    let integer = sqlite3_column_int(statement.pointer, i)
-                    row.data[column] = .integer(Int(integer))
-                    
-                case SQLITE_FLOAT: // as in floating number, actually returns a double.
-                    let double = Double(sqlite3_column_double(statement.pointer, i))
-                    row.data[column] = .double(double)
-                case SQLITE_NULL:
-                    row.data[column] = .null
-                    
-                default:
-                    throw SQLiteError.execute("unsupported type")
-                }
-
+                try row.bind(at: i, pointer: statement.pointer)
             }
 
             result.rows.append(row)
@@ -160,34 +126,6 @@ public class SQLite {
     }
 }
 
-extension SQLite {
-    /**
-        Represents a row of data from
-        a SQLite table.
-    */
-    public struct Result {
-        
-        public struct Row {
-            public var data: [String: DataType]
-
-            init() {
-                data = [:]
-            }
-            
-            public subscript(column: String) -> DataType {
-                return data[column]!
-            }
-            
-        }
-
-        var rows: [Row]
-
-        init() {
-            rows = []
-        }
-    }
-}
-
 extension SQLite.Database {
     /**
         Returns the last error message
@@ -203,96 +141,3 @@ extension SQLite.Database {
 
 }
 
-extension SQLite {
-    /**
-        Represents a single database statement.
-        The statement is used to bind prepared
-        values and contains a pointer to the 
-        underlying SQLite statement memory.
-    */
-    public class Statement {
-        public typealias Pointer = OpaquePointer
-
-        public var pointer: Pointer
-        public var database: Database
-
-        var bindPosition: Int32
-        var nextBindPosition: Int32 {
-            bindPosition += 1
-            return bindPosition
-        }
-
-        public init(pointer: Pointer, database: Database) {
-            self.pointer = pointer
-            self.database = database
-            bindPosition = 0
-        }
-
-        public func reset(_ statementPointer: OpaquePointer) {
-            sqlite3_reset(statementPointer)
-            sqlite3_clear_bindings(statementPointer)
-        }
-
-        public func bind(_ value: Double) throws {
-            if sqlite3_bind_double(pointer, nextBindPosition, value) != SQLITE_OK {
-                throw SQLiteError.bind(database.errorMessage)
-            }
-        }
-
-        public func bind(_ value: Int) throws {
-            if sqlite3_bind_int(pointer, nextBindPosition, Int32(value)) != SQLITE_OK {
-                throw SQLiteError.bind(database.errorMessage)
-            }
-        }
-
-        public func bind(_ value: String) throws {
-            let strlen = Int32(value.utf8.count)
-            if sqlite3_bind_text(pointer, nextBindPosition, value, strlen, SQLITE_TRANSIENT) != SQLITE_OK {
-                throw SQLiteError.bind(database.errorMessage)
-            }
-        }
-
-        public func bind(_ value: Bool) throws {
-            try bind(value ? 1 : 0)
-        }
-        
-
-        public func null()  throws {
-            if sqlite3_bind_null(pointer, nextBindPosition) != SQLITE_OK {
-                throw SQLiteError.bind(database.errorMessage)
-            }
-        }
-    }
-}
-
-extension SQLite {
-    /**
-     enumerates all possible SQLite datatypes :
-     - integer
-     - text
-     - double
-     - null
-     */
-    public enum DataType: Equatable {
-        case integer(Int)
-        case text(String)
-        case double(Double)
-        case null
-        static public func ==(lhs: SQLite.DataType, rhs: SQLite.DataType) -> Bool {
-            switch (lhs, rhs) {
-            case (.double(let double1), .double(let double2)):
-                return double1 == double2
-            case (.text(let string1), .text(let string2)):
-                return string1 == string2
-            case (.integer(let int1), .integer(let int2)):
-                return int1 == int2
-            case (.null, .null):
-                return true
-            default:
-                return false
-            }
-        }
-    }
-    
-    
-}
